Add original Device Tree Source from Armbian.
This is completely default DTS included in
Armbian image. Just uploading as a base for
changes.

Device Tree Blob made from this DTS is not used
by default for Banana Pi M1+ though, it needs
to be chosen in /boot/armbianEnv.txt. It will make
the Broadcom Wifi driver work.